+++ src/Security/Security/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.7.8
 * Add new cmdlets: `Get-AzRegulatoryComplianceStandard`, 
 					`Get-AzRegulatoryComplianceControl`, 
 					`Get-AzRegulatoryComplainceAssessment`